get_caged/pyproject.toml
[tool.poetry]
name = "get_caged"
version = "0.1.0"
description = ""
authors = ["Helena Strandberg <hejhej.helena@gmail.com>", "Alexander Hultnér <ahultner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.2"
pydantic = "^1.5.1"
pillow = "^7.1.2"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
bpython = "^0.19"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
